requirements.txt
fastapi
uvicorn
pydantic
pandas
lightgbm
scikit-learn
joblib
python-dotenv
imbalanced-learn